COMET: Added debug console commands to enable debug test and cheating with block puzzle.
The "puzzleCheat" and "testPuzzle" commands at the debug console replace compile time selection of these functions.

If it is desirable to have these activated at startup for development (for those with an aversion to the ESC, CTRL-D keys ;-), then it should be very easy to set the debug booleans in console.cpp by a bootflag.

Also, the "puzzleCheat" command or similar should be retained into the stable engine to ensure frustration-free playtesting.

git-svn-id: https://svn.scummvm.org:4444/svn/comet@209 e395a8f9-5cf0-422e-971d-9a083d8f24e8