COMET: Refactor a lot of code to use a task/async-based approach
This is intended to solve several problems:
- Events are handled as soon as possible:
  Input state doesn't have to be remembered when the event occurs and recalled
  later which could cause events to be skipped
- Game update/framerate is controlled in a single place:
  syncUpdate (i.e. delayMillis) isn't done all over the place any more
  This also avoids having to test shouldQuit in many places to break out of
  loops or return from functions, which simplifies the code
- Smoother mouse cursor since delayMillis isn't called with long delays during
  which no screen updates were performed, resulting in choppy mouse cursor
  movement.
- Bonus side effect of extracting various methods into Task-classes

The task system is modeled similarly to the coroutine mechanism used in some
games but less complex since only a single task can be active at a time.
Instead of hand-coding state machines it uses the same concept of using the
__LINE__ macro together with a switch statement to go the next instruction.
This is wrapped inside some simple macros.

A task is a class derived from the TaskBase class and overrides some methods.
A task can either call a task-enabled method or another task. The calling task
is then paused, waiting for the called method/task to finish.
The task body is executed repeatedly until explicitly terminated, while a
method is run only once. Tasks and methods can yield to return control to the
main loop which in turn handles event polling and delegation, quit requests and
framerate control.

The current code still needs some work, e.g. some objects are currently never
freed and macros for tasks and methods can probably merged.