TTS: MACOS, IOS: Implement Text to Speech using AVSpeechSynthesizer
The NSSpeechSynthesizer API has been deprecated in recent macOS versions.
This commits adds an implementation using the newer AVSpeechSynthesizer,
available since macOS 10.14. The new implementation is only used when
building for Apple Silicon. In theory we could also use it when building
for 64 bit Intel macs and targetting macOS 10.14 or above, but there is
not much benefit in doing so.

As a bonus it should also work on iOS.